feat: migrate money storage to integer paise
Squash merge of the verified minor-unit money migration. Required Android CI and dependency review checks passed.

+++ app/src/main/java/com/example/ui/screens/HomeScreen.kt
@@ -471,7 +471,7 @@ fun HomeScreen(viewModel: ShopViewModel) {
 @Composable
 fun MetricSummaryTile(
     title: String,
-    amount: Double,
+    amount: Long,
     accentColor: Color,
     containerColor: Color,
     icon: ImageVector,
